clientSideApp/requirements.txt

# This file is part of the client-side application requirements.
Flask==2.3.3
requests==2.31.0
pymongo==4.7.2
python-dotenv
dnspython